+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- PR [173](https://github.com/marinasundstrom/CheckedExceptions/pull/173) Handle inheritance in redundancy check
+
 ## [1.7.1] - 2025-07-30
 
 ### Fixed