+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - PR [173](https://github.com/marinasundstrom/CheckedExceptions/pull/173) Handle inheritance in redundancy check
 
+- PR [#179](https://github.com/marinasundstrom/CheckedExceptions/pull/179) Fix display format of exception names in diagnostics
+
 ## [1.7.1] - 2025-07-30
 
 ### Fixed