+++ CHANGELOG.md
@@ -10,7 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- PR [177](https://github.com/marinasundstrom/CheckedExceptions/pull/177) Handle property decl in XML doc analyser 
+- PR [#177](https://github.com/marinasundstrom/CheckedExceptions/pull/177) Handle property decl in XML doc analyser
+
+- PR [#183](https://github.com/marinasundstrom/CheckedExceptions/pull/177) Ability to turn off XML doc interop feature
 
 ### Fixed